requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
torch==2.1.0+cu118
torchvision==0.16.0+cu118
torchaudio==2.1.0+cu118
diffusers==0.21.4
transformers==4.30.2
accelerate==0.20.3
safetensors==0.3.1
xformers==0.0.20
opencv-python==4.8.0.76
Pillow==9.5.0
numpy==1.24.3
scipy==1.10.1
pydantic==2.5.0
python-multipart==0.0.6
